IoT_Core/requirements.txt
Django>=5.1, <5.3
Celery
redis
django-celery-beat
django-celery-results
dj-database-url
psycopg[binary]
python-decouple
django-timescaledb
gunicorn
kombu
whitenoise
djangorestframework
